requirements.txt
tk 0.1.0
Pillow 9.2.0
matplotlib 3.6.0
pandas 1.5.0
typing 3.7.4.3
pyenchant 3.2.2
